TVGUI-13 Move AddShow component into folder
